number-detection/requirements.txt
tensorflow
numpy
mnist
keras
fastapi
uvicorn 
python-multipart
Pillow